refactor(ui): split footer types and fix hook lint errors
