+++ src/pages/home/ui.tsx
@@ -1,8 +1,7 @@
 import { useHomeLogic } from "./logic";
 import HushhTechHeader from "../../components/hushh-tech-header/HushhTechHeader";
-import HushhTechFooter, {
-  HushhFooterTab,
-} from "../../components/hushh-tech-footer/HushhTechFooter";
+import HushhTechFooter from "../../components/hushh-tech-footer/HushhTechFooter";
+import { HushhFooterTab } from "../../components/hushh-tech-footer/types";
 import HushhTechCta, {
   HushhTechCtaVariant,
 } from "../../components/hushh-tech-cta/HushhTechCta";
