+++ src/pages/hushh-user-profile/ui.tsx
@@ -11,7 +11,8 @@ import { FaApple } from "react-icons/fa";
 import { FcGoogle } from "react-icons/fc";
 import HushhTechBackHeader from "../../components/hushh-tech-back-header/HushhTechBackHeader";
 import HushhTechCta, { HushhTechCtaVariant } from "../../components/hushh-tech-cta/HushhTechCta";
-import HushhTechFooter, { HushhFooterTab } from "../../components/hushh-tech-footer/HushhTechFooter";
+import HushhTechFooter from "../../components/hushh-tech-footer/HushhTechFooter";
+import { HushhFooterTab } from "../../components/hushh-tech-footer/types";
 import NWSScoreBadge from "../../components/profile/NWSScoreBadge";
 import { PrivacyShield } from "../../components/profile/PrivacyShield";
 import WalletCardPreviewModal from "../../components/wallet/WalletCardPreviewModal";
