+++ src/pages/profile/ui.tsx
@@ -7,7 +7,8 @@
 import React from 'react';
 import HushhTechBackHeader from '../../components/hushh-tech-back-header/HushhTechBackHeader';
 import HushhTechCta, { HushhTechCtaVariant } from '../../components/hushh-tech-cta/HushhTechCta';
-import HushhTechFooter, { HushhFooterTab } from '../../components/hushh-tech-footer/HushhTechFooter';
+import HushhTechFooter from '../../components/hushh-tech-footer/HushhTechFooter';
+import { HushhFooterTab } from '../../components/hushh-tech-footer/types';
 import { useProfileLogic } from './logic';
 
 /* ── Playfair heading style ── */
